PF-132: fix quest score-gain JSON serialization; harden numeric-accuracy eval
- JourneyAdvisorClient now accepts numeric fields serialized as JSON strings
  (a Pydantic Decimal quirk) via JsonNumberHandling.AllowReadingFromString;
  added regression tests covering both the string and number encodings.
- Quest.estimated_score_gain switched from Decimal to float on the Python
  side so it serializes as a JSON number, not a string, in the first place.
- eval_numeric_accuracy.py: fixed a UnicodeEncodeError crash on Windows
  consoles (cp1252 can't render the checkmark markers) and added request
  pacing to stay under the Gemini free-tier rate limit.
- .env.example: added inline "options:" comments documenting valid values
  for every config key, as a quick-reference guide.
- PF-AI008 (Chapter 8 learning plan): recorded a real conversation-memory
  gap found during live testing ("bulan itu" not resolving to the prior
  turn's month) as a concrete motivating example.

+++ services/ai-service/app/models.py
@@ -103,7 +103,10 @@ class Quest(BaseModel):
     title: str
     description: str
     target_indicator: str
-    estimated_score_gain: Decimal
+    # Score gain in points (0–100), not a monetary value — must serialize as a
+    # JSON number. Pydantic renders Decimal as a JSON *string* ("12"), which the
+    # .NET client cannot read into `decimal`; float serializes as a number.
+    estimated_score_gain: float
     difficulty: Literal["easy", "medium", "hard"]
     action_deeplink: str | None = None
